add django rest know
Knox tokens should be generated using the provided views. Any APIView or
ViewSet can be accessed using these tokens by adding TokenAuthentication
to the View's authentication_classes. To authenticate, the Authorization
header should be set on the request, with a value of the word "Token",
then a space, then the authentication token provided by LoginView.